Correction to display_pylons
It is legal to place the small disk on top of the large disk. However as it was before it would not display the small disk if it was above the large disk. This line will now display the small disk above the large one if the medium disk is not present.